dom/contacts/tests/test_contacts_substringmatchingVE.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=877302
-->
<head>
  <title>Test for Bug 877302 substring matching for WebContacts</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=877302">Mozilla Bug 877302</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="text/javascript;version=1.8" src="http://mochi.test:8888/tests/dom/contacts/tests/shared.js"></script>
<script class="testbody" type="text/javascript">
"use strict";

var substringLength = 7;
SpecialPowers.setIntPref("dom.phonenumber.substringmatching.VE", substringLength);
SpecialPowers.setCharPref("ril.lastKnownSimMcc", "734");

var sample_id1;
var createResult1;
var findResult1;

var prop = {
  tel: [{value: "7932012345" }, {value: "7704143727591"}]
};

var prop2 = {
  tel: [{value: "7932012345" }, {value: "+58 212 5551212"}]
};

var req;
var steps = [
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear()
    req.onsuccess = function () {
      ok(true, "Deleted the database");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding contact");
    createResult1 = new mozContact(prop);
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find({});
    req.onsuccess = function () {
      is(req.result.length, 1, "One contact.");
      findResult1 = req.result[0];
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 1");
    var length = prop.tel[0].value.length;
    var num = "04143727591"
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: num};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      is(findResult1.tel[1].value, "7704143727591", "Same Value");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding contact");
    createResult1 = new mozContact(prop2);
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 2");
    var num = "5551212";
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: num};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      is(findResult1.tel[1].value, "+58 212 5551212", "Same Value");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.clear()
    req.onsuccess = function () {
      ok(true, "Deleted the database");
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "all done!\n");
    SpecialPowers.setIntPref("dom.phonenumber.substringmatching.VE", -1);
    SimpleTest.finish();
  }
];

start_tests();
</script>
</pre>
</body>
</html>
